feat(client): add address CRUD features

+++ P01-pojo/src/main/java/com/fallensakura/entity/Address.java
@@ -59,5 +59,5 @@ public class Address implements Serializable {
      * 1是 0否
      */
     @Schema(description = "1是 0否")
-    private Byte isDefault;
+    private Integer isDefault;
 }
